Add documentation for vector & move it to new pkg
