+++ pkg/primitives/vector/vector.go
@@ -1,8 +1,7 @@
-package primitives
+package vector
 
 import (
 	"math"
-	"math/rand"
 	"strconv"
 )
 
@@ -12,16 +11,6 @@ type Vector struct {
 
 var UnitVector = Vector{1, 1, 1}
 
-func VectorInUnitSphere(rnd *rand.Rand) Vector {
-	for {
-		r := Vector{rnd.Float64(), rnd.Float64(), rnd.Float64()}
-		p := r.MultiplyScalar(2.0).Subtract(UnitVector)
-		if p.SquaredLength() >= 1.0 {
-			return p
-		}
-	}
-}
-
 // ToString returns the string representation of the vector
 func (v Vector) ToString() string {
 	x := strconv.FormatFloat(v.X, 'f', 6, 64)
@@ -29,22 +18,23 @@ func (v Vector) ToString() string {
 	z := strconv.FormatFloat(v.Z, 'f', 6, 64)
 	return "X : " + x + ", Y : " + y + ", Z : " + z
 }
+
+// Length returns the length of the vector.
 func (v Vector) Length() float64 {
 	return math.Sqrt(v.X*v.X + v.Y*v.Y + v.Z*v.Z)
 }
 
-func (v Vector) SquaredLength() float64 {
-	return v.X*v.X + v.Y*v.Y + v.Z*v.Z
-}
-
+// Normalize returns the normalized vector. This operand makes the length of the vector 1.
 func (v Vector) Normalize() Vector {
 	return v.DivideScalar(v.Length())
 }
 
+// Dot aka SquaredLength returns the squared length of the vector. (length^2)
 func (v Vector) Dot(ov Vector) float64 {
 	return v.X*ov.X + v.Y*ov.Y + v.Z*ov.Z
 }
 
+// Cross returns the cross product vector of the given vectors.
 func (v Vector) Cross(ov Vector) Vector {
 	return Vector{
 		v.Y*ov.Z - v.Z*ov.Y,
@@ -53,38 +43,51 @@ func (v Vector) Cross(ov Vector) Vector {
 	}
 }
 
+// Add returns the sum of the 2 vectors.
 func (v Vector) Add(ov Vector) Vector {
 	return Vector{v.X + ov.X, v.Y + ov.Y, v.Z + ov.Z}
 }
 
+// Subtract returns the substracted vector.
 func (v Vector) Subtract(ov Vector) Vector {
 	return Vector{v.X - ov.X, v.Y - ov.Y, v.Z - ov.Z}
 }
 
+// Multiply returns the multiplied vector.
 func (v Vector) Multiply(ov Vector) Vector {
 	return Vector{v.X * ov.X, v.Y * ov.Y, v.Z * ov.Z}
 }
 
+// Devide returns the divided vector.
 func (v Vector) Divide(ov Vector) Vector {
 	return Vector{v.X / ov.X, v.Y / ov.Y, v.Z / ov.Z}
 }
 
+// AddScalar returns the incremented vector.
+// v.AddScalar(t) = v.Add(vector{t,t,t})
 func (v Vector) AddScalar(t float64) Vector {
 	return Vector{v.X + t, v.Y + t, v.Z + t}
 }
 
+// SubtractScalar returns the decremented vector.
+// v.SubtractScalar(t) = v.Subtract(vector{t,t,t})
 func (v Vector) SubtractScalar(t float64) Vector {
 	return Vector{v.X - t, v.Y - t, v.Z - t}
 }
 
+// MultiplyScalar returns the multiplied vector.
+// v.MultiplyScalar(t) = v.Multiply(vector{t,t,t})
 func (v Vector) MultiplyScalar(t float64) Vector {
 	return Vector{v.X * t, v.Y * t, v.Z * t}
 }
 
+// DivideScalar returns the divided vector.
+// v.DivideScalar(t) = v.Divide(vector{t,t,t})
 func (v Vector) DivideScalar(t float64) Vector {
 	return Vector{v.X / t, v.Y / t, v.Z / t}
 }
 
+/*
 func (v Vector) Reflect(ov Vector) Vector {
 	b := 2 * v.Dot(ov)
 	return v.Subtract(ov.MultiplyScalar(b))
@@ -102,3 +105,14 @@ func (v Vector) Refract(ov Vector, n float64) (bool, Vector) {
 	}
 	return false, Vector{}
 }
+
+func VectorInUnitSphere(rnd *rand.Rand) Vector {
+	for {
+		r := Vector{rnd.Float64(), rnd.Float64(), rnd.Float64()}
+		p := r.MultiplyScalar(2.0).Subtract(UnitVector)
+		if p.Dot() >= 1.0 {
+			return p
+		}
+	}
+}
+*/
