+++ pkg/primitives/vector/vector_test.go
@@ -1,4 +1,4 @@
-package primitives
+package vector
 
 import (
 	"math"
